# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Cloudflare Backend Boilerplate API",
		"version": "0.1.0",
		"description": "API-only Cloudflare Worker (Hono) for mobile clients. Generated from src/routes/* and src/lib/schemas.ts — see scripts/generate-openapi.ts."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:8787",
			"description": "Local dev"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Better Auth JWT access token (15m). Also accepts an opaque session/API token on some routes — see docs/README."
			}
		}
	},
...
	"paths": {
...
		"/api/v1/users/me": {
			"get": {
				"tags": [
					"Users"
				],
				"summary": "Get the current user's profile",
				"responses": {
					"200": {
						"description": "User profile",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"patch": {
				"tags": [
					"Users"
				],
				"summary": "Update the current user's profile",
				"responses": {
					"200": {
						"description": "Updated user profile",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"minLength": 1,
										"maxLength": 120
									},
									"avatar": {
										"type": [
											"string",
											"null"
										]
									},
									"bio": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 500
											},
											{
												"type": "null"
											}
										]
									},
									"company": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 120
											},
											{
												"type": "null"
											}
										]
									},
									"jobTitle": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 120
											},
											{
												"type": "null"
											}
										]
									},
									"phone": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 40
											},
											{
												"type": "null"
											}
										]
									},
									"website": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 200
											},
											{
												"type": "null"
											}
										]
									},
									"country": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 80
											},
											{
												"type": "null"
											}
										]
									},
									"timezone": {
										"anyOf": [
											{
												"type": "string",
												"maxLength": 64
											},
											{
												"type": "null"
											}
										]
									},
									"onboardingCompleted": {
										"type": "boolean"
									},
									"onboardingStep": {
										"type": "integer",
										"minimum": 0,
										"maximum": 100
									}
								},
								"additionalProperties": false
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Users"
				],
				"summary": "Soft-delete the current user's account",
				"responses": {
					"200": {
						"description": "Account deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/users/me/change-password": {
			"post": {
				"tags": [
					"Users"
				],
				"summary": "Change the current user's password",
				"responses": {
					"200": {
						"description": "Password updated",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"currentPassword": {
										"type": "string",
										"minLength": 1
									},
									"newPassword": {
										"type": "string",
										"minLength": 8
									}
								},
								"required": [
									"currentPassword",
									"newPassword"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/users/me/avatar": {
			"post": {
				"tags": [
					"Users"
				],
				"summary": "Upload a profile avatar (multipart/form-data, field `file`)",
				"responses": {
					"200": {
						"description": "Updated user profile with new avatar URL",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": [
					"Users"
				],
				"summary": "Remove the current user's avatar",
				"responses": {
					"200": {
						"description": "Avatar removed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/users/me/devices": {
			"post": {
				"tags": [
					"Users"
				],
				"summary": "Register a push notification device token",
				"responses": {
					"200": {
						"description": "Device token registered or updated",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"token": {
										"type": "string",
										"minLength": 1
									},
									"platform": {
										"type": "string",
										"enum": [
											"ios",
											"android"
										]
									}
								},
								"required": [
									"token",
									"platform"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/users/me/devices/{token}": {
			"delete": {
				"tags": [
					"Users"
				],
				"summary": "Unregister a push notification device token",
				"responses": {
					"200": {
						"description": "Device token removed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/uploads": {
			"post": {
				"tags": [
					"Uploads"
				],
				"summary": "Upload a file to R2 (multipart/form-data field `file`, or JSON base64)",
				"responses": {
					"200": {
						"description": "Stored object key, URL (if public), size and MIME type",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"fileBase64": {
										"type": "string",
										"minLength": 1
									},
									"mimeType": {
										"type": "string",
										"minLength": 1
									},
									"filename": {
										"type": "string"
									}
								},
								"required": [
									"fileBase64",
									"mimeType"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
...
	}
}